1. move tracer from service folder to search folder. 2. modify next/copy a bit for rest individual and one max individual.
